# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Verified B2B Deal Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: Discover organizations with public stats
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search legal/trading names
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: industry
          schema:
            type: string
        - in: query
          name: verificationStatus
          schema:
            type: string
            enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED, EXPIRED]
        - in: query
          name: verifiedOnly
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Organization discovery results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationDiscoveryItem"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization profile
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}/stats:
    get:
      operationId: getOrganizationStats
      tags: [organizations]
      summary: Get public organization stats and reputation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Public organization stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDiscoveryItem"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations/{orgId}/members:
    get:
      operationId: listOrganizationMembers
      tags: [organizations]
      summary: List organization members
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationMember"
    post:
      operationId: inviteOrganizationMember
      tags: [organizations]
      summary: Invite a member to the organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInviteInput"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMember"
  /organizations/{orgId}/members/{memberId}:
    patch:
      operationId: updateOrganizationMember
      tags: [organizations]
      summary: Update member role or signer status
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMember"
    delete:
      operationId: removeOrganizationMember
      tags: [organizations]
      summary: Remove a member from the organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── Verification ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ────────────────────────────────────────────────────────────────────
...
    OrganizationUpdate:
      type: object
      additionalProperties: false
      properties:
        tradingName:
          type: ["string", "null"]
        regionsServed:
          type: array
          items:
            type: string
        publicProfile: {}
    Organization:
      type: object
      additionalProperties: false
      required: [id, legalName, country, industry, verificationStatus, createdAt]
      properties:
        id:
          type: string
        legalName:
          type: string
        tradingName:
          type: ["string", "null"]
        country:
          type: string
        industry:
          type: string
        regionsServed:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED, EXPIRED]
        publicProfile: {}
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReputationSummary:
      type: object
      additionalProperties: false
      required: [completedDeals, disputedDeals, financedDeals, totalVolume, score]
      properties:
        completedDeals:
          type: string
        disputedDeals:
          type: string
        financedDeals:
          type: string
        totalVolume:
          type: string
        score:
          type: string
    OrganizationDiscoveryStats:
      type: object
      additionalProperties: false
      required: [identityId, verifiedAt, verificationExpiresAt, reputation]
      properties:
        identityId:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        verificationExpiresAt:
          type: ["string", "null"]
          format: date-time
        reputation:
          $ref: "#/components/schemas/ReputationSummary"
    OrganizationDiscoveryItem:
      type: object
      additionalProperties: false
      required: [id, legalName, country, industry, verificationStatus, createdAt, stats]
      properties:
        id:
          type: string
        legalName:
          type: string
        tradingName:
          type: ["string", "null"]
        country:
          type: string
        industry:
          type: string
        regionsServed:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED, EXPIRED]
        publicProfile: {}
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        stats:
          $ref: "#/components/schemas/OrganizationDiscoveryStats"
    OrganizationMember:
      type: object
      additionalProperties: false
      required: [id, organizationId, userId, role, isAuthorizedSigner, joinedAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [OWNER, ADMIN, FINANCE, MEMBER]
        isAuthorizedSigner:
          type: boolean
        invitedBy:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time
        revokedAt:
          type: ["string", "null"]
    MemberInviteInput:
      type: object
      additionalProperties: false
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, FINANCE, MEMBER]
        isAuthorizedSigner:
          type: boolean
    MemberUpdate:
      type: object
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [ADMIN, FINANCE, MEMBER]
        isAuthorizedSigner:
          type: boolean

    # ── Verification ────────────────────────────────────────────────────────────
...